+++ db/migrate/20180123205423_add_report_manager_to_users.rb
@@ -1,0 +1,6 @@
+class AddReportManagerToUsers < ActiveRecord::Migration[5.2]
+  def change
+    add_column :users, :report_manager, :boolean, null: false, default: false
+    add_index :users, :report_manager
+  end
+end